pyproject.toml
[project]
name = "filadaminha"
version = "0.1.0"
description = "A simple API to manage a queue of people"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.5",
    "pydantic[email]>=2.9.2",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.17",
    "qrcode>=8.0",
    "uvicorn>=0.32.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "ruff>=0.8.0",
]

